+++ context.go
@@ -17,7 +17,7 @@ import (
 // Version of GoatCounter; set at compile-time with:
 //
 //	-ldflags="-X zgo.at/goatcounter/v2.Version=…"
-var Version = "dev"
+var Version = "v3.1.2-lite"
 
 func getCommit() (string, time.Time, bool) {
 	var (